extract generateExtrinsicColumns helper
