+++ client/src/Block/components/BlockDetailsExtrinsicList.tsx
@@ -2,12 +2,10 @@ import { FC } from 'react'
 import { Extrinsic } from 'gql/graphql'
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
-import { Link } from 'react-router-dom'
 
 // common
-import { Table, Column } from 'common/components'
-import { shortString } from 'common/helpers'
-import { INTERNAL_ROUTES } from 'common/routes'
+import { Table } from 'common/components'
+import { generateExtrinsicColumns } from 'common/helpers/generateColumns'
 
 dayjs.extend(relativeTime)
 
@@ -16,44 +14,7 @@ type Props = {
 }
 
 const BlockDetailsExtrinsicList: FC<Props> = ({ extrinsics }) => {
-  // methods
-  const generateColumns = (extrinsics: Extrinsic[]): Column[] => [
-    {
-      title: 'Extrinsic Id',
-      cells: extrinsics.map(({ block, pos, id }) => (
-        <Link key={`${id}-block-extrinsic-id`} to={INTERNAL_ROUTES.extrinsics.id.page(id)}>
-          <div>{`${block.height}-${pos}`}</div>
-        </Link>
-      )),
-    },
-    {
-      title: 'Block hash',
-      cells: extrinsics.map(({ hash, id }) => (
-        <div key={`${id}-block-extrinsic-hash`}>{shortString(hash)}</div>
-      )),
-    },
-    {
-      title: 'Action',
-      cells: extrinsics.map(({ name, id }) => (
-        <div key={`${id}-block-extrinsic-action`}>{name.split('.')[1].toUpperCase()}</div>
-      )),
-    },
-    {
-      title: 'Time',
-      cells: extrinsics.map(({ block, id }) => {
-        const blockDate = dayjs(block.timestamp).fromNow(true)
-
-        return <div key={`${id}-block-extrinsic-time`}>{blockDate}</div>
-      }),
-    },
-    {
-      title: 'Status',
-      cells: extrinsics.map(() => <></>),
-    },
-  ]
-
-  // constants
-  const columns = generateColumns(extrinsics)
+  const columns = generateExtrinsicColumns(extrinsics)
 
   return (
     <Table
